+++ src/frontend/app/app.common.module.ts
@@ -1,6 +1,6 @@
 import {AppComponent} from './components/app/app.component';
 import {AppService} from './services/app.service';
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import {NgbCollapseModule} from '@ng-bootstrap/ng-bootstrap';
 import {NavbarComponent} from './components/navbar/navbar.component';
 import {ToastrModule} from 'ngx-toastr';
 import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
@@ -58,7 +58,7 @@ export function getAngulartics2Providers(): any {
     ],
     imports: [
         HttpClientModule,
-        NgbModule,
+        NgbCollapseModule,
         BrowserModule.withServerTransition({appId: 'pdoc-app'}),
         ToastrModule.forRoot(),
         TranslateModule.forRoot({