+++ src/frontend/environments/app-environment.ts
@@ -2,6 +2,9 @@ import {CommonEnvironment} from '@dps/mycms-frontend-commons/dist/frontend-pdoc-
 import {DataMode} from '../shared/commons/model/datamode.enum';
 
 export interface AppEnvironment extends CommonEnvironment {
+    hideCopyrightFooter?: boolean,
+    hideInternalDescLinks?: boolean;
+    hideInternalImages?: boolean,
     assetsPathVersionSnippet: string;
     assetsPathVersionSuffix: string;
     currentDataMode?: DataMode;