+++ src/frontend/environments/environment.viewer.ts
@@ -2,6 +2,7 @@ import {AppEnvironment} from './app-environment';
 import {DataMode} from '../shared/commons/model/datamode.enum';
 
 export const environment: AppEnvironment = {
+    hideCopyrightFooter: true,
     production: false, // TODO set this to false for development
     assetsPathVersionSnippet: '',
     assetsPathVersionSuffix: '',
@@ -13,6 +14,8 @@ export const environment: AppEnvironment = {
     cookieLawSeenName: 'cookieLawSeenV20180525',
     trackingProviders: [], // Angulartics2Piwik
     staticPDocsFile: 'assets/staticdata/static.myshppdocs.js',
+    hideInternalDescLinks: true,
+    hideInternalImages: true,
     startDataMode: DataMode.STATIC,
     availableDataModes: [DataMode.STATIC]
 };